requirements.txt
ddddocr
python-socketio[client]
loguru
openai
requests
pillow
pytz
aiohttp
langchain
langchain_core
langchain_openai
langchain_community
pandas
pydantic
requests
faiss-cpu
openpyxl